+++ content/template/detailTemplate/template_A.php
@@ -41,7 +41,7 @@ else
 	if(!$selectedCourse)
 	{
 		?>
-		<script>history.go(-1);</script>
+		<script type="text/javascript">location.href = '<?php echo $baseUrl; ?>';</script>
 		<?php
 		die();
 	}